+++ images/syncer/Dockerfile
@@ -49,6 +49,6 @@ COPY --from=builder /build/vsphere-syncer /bin/vsphere-syncer
 
 RUN mkdir -p config
 
-ADD pkg/syncer/cnsoperator/config/cnsregistervolume_crd.yaml /config/
+ADD pkg/apis/cnsoperator/config/cnsregistervolume_crd.yaml /config/
 
 ENTRYPOINT ["/bin/vsphere-syncer"]
