+++ pkg/apis/cnsoperator/cnsnodevmattachment/v1alpha1/cnsnodevmattachment_types.go
