+++ pkg/apis/cnsoperator/cnsnodevmattachment/v1alpha1/zz_generated.deepcopy.go
