+++ pkg/apis/cnsoperator/config/cnsregistervolume_crd.yaml
