+++ pkg/apis/storagepool/config/cns.vmware.com_storagepools.yaml
