+++ pkg/apis/storagepool/config/sample_v1alpha1_storagepool.yaml
